+++ inc/staffSvg/staffNotation.ts
@@ -415,9 +415,9 @@ const parseNotationInMeasure = (context: StaffContext, measure) => {
 				context.setBeatsPerMeasure(token.timeSignatureValue);
 		}
 		else if (token.is("NOTEHEAD")) {
-			// ignore tempo note heads
+			/*// ignore tempo note heads
 			if (token.source.substr(0, 6) === "\\tempo")
-				continue;
+				continue;*/
 
 			const contextIndex = context.snapshot();