+++ backend/scoreMaker.ts
@@ -18,11 +18,9 @@ interface LilyProcessOptions {};
 const xmlToLyWithMarkup = async (xml: Buffer, options: LilyProcessOptions, markup: string): Promise<string> => {
 	const bom = (xml[0] << 8 | xml[1]);
 	const utf16 = bom === 0xfffe;
-	let content = xml.toString(utf16 ? "utf16le" : "utf8");
-	if (utf16)
-		content = content.replace(/encoding='UTF-16'/i, "encoding='UTF-8'");
+	const content = xml.toString(utf16 ? "utf16le" : "utf8");
 
-	const lily = await xml2ly(content, options);
+	const lily = await xml2ly(content, {replaceEncoding: utf16, ...options});
 
 	// copy markup
 	if (markup) {